change size_t to int(for now)

+++ stacks/stack.h
@@ -1,7 +1,7 @@
 typedef struct Stack {
   int top;
   int *data;
-  size_t capacity;
+  int capacity;
 
 } Stack;
